Créer une interface pour permettre de saisir un nouvel utilisateur Fixes #47
